requirements.txt
click
python-dotenv
spotipy
ytmusicapi

pre-commit
